// dashboard/src/utils/hasura-api/generated/schemas/createArrayRelationshipArgs.ts
/**
 * Generated by orval v7.18.0 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { QualifiedTable } from './qualifiedTable';
import type { RelationshipUsing } from './relationshipUsing';

export interface CreateArrayRelationshipArgs {
  /** Name of the table */
  table: QualifiedTable;
  /** Name of the new relationship */
  name: string;
  /** Use one of the available ways to define an array relationship */
  using: RelationshipUsing;
  comment?: string;
  /** Name of the source database of the table */
  source?: string;
}